Create Virtual-function.cpp
This program demonstrates the use of virtual functions. What is a virtual function? A virtual function is a member function that is declared within a base class and is redefined (overridden) by a derived class. When you refer to a derived class object using a pointer or a reference to the base class, you can call a virtual function for that object and execute the derived class’s version of the method.

This program will display information about three people. Common attributes like ID, Name, and Date of Birth will be defined in the virtual function, whereas, when defining a class for each person, we will add separate attributes that will be included with the virtual function.